[#27] Revert COOLDOWN_PERIOD to uint64

+++ contracts/SafeTokenLock.sol
@@ -25,7 +25,7 @@ contract SafeTokenLock is ISafeTokenLock, IRecoverERC20, Ownable2Step {
 
     /* solhint-disable var-name-mixedcase */
     IERC20 public immutable SAFE_TOKEN; // Safe Token Address.
-    uint32 public immutable COOLDOWN_PERIOD; // Contains the cooldown period. Default will be 30 days.
+    uint64 public immutable COOLDOWN_PERIOD; // Contains the cooldown period. Default will be 30 days.
 
     /* solhint-enable var-name-mixedcase */
     mapping(address => User) public users; // Contains the address => user info struct.
